v0.15.0:#75: Adds default error options to targets.
Adds the ability to specify default error options for targets.

This change introduces a new function `target_default_error_options` that applies error-related compiler flags. The CMake files for executables, libraries, and tests are updated to include an option to enable these default error options.
This provides a way to enforce stricter compilation and catch potential issues early on.